// src/entities/openapi/model/targetTypeCodeEnum.ts
/**
 * agridata.ch API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const TargetTypeCodeEnum = {
    DataRequest: 'DATA_REQUEST'
} as const;
export type TargetTypeCodeEnum = typeof TargetTypeCodeEnum[keyof typeof TargetTypeCodeEnum];